% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars.R
\name{bars}
\alias{bars}
\title{Create Bar Charts}
\usage{
bars(data, count, group, highlight_variable = NA,
  ordered = "alphabetically", first_group = NA, flipped = TRUE,
  labelled = FALSE, show_percent = NA, label_nudge_y = -(max(data
  \%>\% select({     {         count     } }))/100 * 2),
  label_nudge_x = 0, label_size = 3, label_color = "white",
  label_hjust = "right", axis.title.x = element_blank(),
  axis.title.y = element_blank(), title = element_blank(),
  legend.position = "none", fontsize = 16)
}
\arguments{
\item{data}{A data frame.}

\item{count}{The variable that contains the number of observations per group.}

\item{group}{The groups that are displayed as separate bars.}

\item{highlight_variable}{Default NA, indicate variable that is TRUE only for the highlighted group.}

\item{ordered}{Default alphabetically, alternative is count, then ordered in descending order.}

\item{first_group}{Default NA, indicate a group which should be displayed first (left or top).}

\item{flipped}{Default TRUE. If flipped is TRUE bars are horizontally (groups on y axis), otherwise bars are vertically (groups on x axis).}
}
\value{
A ggplot style bar graph.
}
\description{
Create a bar chart based on ggplot.
}
\details{
Use ggplot's geom_col geom to create bar charts that are visually appealing and in line with a common
theme and visualization style.
}
\examples{
\dontrun{
  bars(data = sdg \%>\% count(donor) \%>\% filter(n > 30000) \%>\% mutate(highlight = donor == "Germany"), count = n, group = donor, highlighted_var = highlight, ordered = "count", first_group = "Germany")
  }
}
\author{
Philipp Kollenda
}
