% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emperors.R
\docType{data}
\name{emperors}
\alias{emperors}
\title{Emperor Data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 68 rows and 16 columns.}
\usage{
emperors
}
\description{
A dataset containing information about the life (and death) of Roman emperors.
}
\details{
@format A data frame of 68 rows and 16 columns
 \describe{
  \item{Cause}{Cause of Death}
 }
 @source Somewhere from the internet.
}
\keyword{datasets}
