% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg.R
\docType{data}
\name{sdg}
\alias{sdg}
\title{SDG Data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1252036 rows and 5 columns.}
\usage{
sdg
}
\description{
A dataset containing information on aid projects related to the sustainable development goals.
}
\details{
@format A data frame of 1252036 rows and 5 columns
 \describe{
  \item{ID}{Unique Project ID}
  \item{year}{Year of Project}
  \item{Donor}{Donor Country or Organisation}
  \item{Recipient}{Recipient Country or Organisation}
  \item{Amount}{Total Amount in USD of Project}
 }
 @source Somewhere from the internet.
}
\keyword{datasets}
