% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pkgraph}
\alias{pkgraph}
\alias{pkgraph-package}
\title{pkgraph: Create Beautiful Graphs}
\description{
Easily generate custom graphs from data frames
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pkollenda/pkgraph}
  \item Report bugs at \url{https://github.com/pkollenda/pkgraph/issues}
}

}
\author{
\strong{Maintainer}: Philipp Kollenda \email{p.kollenda@vu.nl}

}
